fix: Return globals even when an error was raised
edxapp expects to read `globals_dict` from the response even when an error
has been caught and signaled with `emsg`. This commit fixes the API
response to match this expectation.

This can't be done without some changes to how we call safe_exec, since we
need to capture both changes to the globals *and* a possible exception.
With the current approach, we can only get one or the other. So to make
this work, we have to take one of two approaches:

A. Stop doing the defensive deepcopy on globals, making the safe_exec
   wrapper a much thinner wrapper than before. The relying code (code-exec
   API and startup checks) would pass in a mutable globals dict and no
   longer expect a return value. When an exception occurs, the mutated
   globals dict would still be available to read.
B. Keep the deepcopy, but move exception-catching into the safe_exec
   wrapper. Return a pair of globals dict and exception (or None). This
   more closely mirrors the actual API we expose in the service. Relying
   code has to destructure the tuple rather than using an except block.

I actually tried approach A at first, but discovered that unit testing
became *far* more unpleasant -- we would have to make a function that
behaves differently each time it's called, but we wouldn't be able to use
the built-in behavior of unittest.mock's side_effects, because it can only
help with return values. And safe_exec would no longer be in the business
of returning things.

With approach B, the mock responses become tuples, which is a little bit
less magic, as we're no longer relying on side_effects' special handling of
exception objects.

Also, update issue link in safe_exec docstring to an older ticket. The
newer one has been closed as a dup.